+++ src/Models/Queries/LocationQuery.php
@@ -12,8 +12,10 @@ class LocationQuery extends I18nActiveQuery
 {
     public function andWhereTagId(int $tagId, bool $eagerLoading = false): static
     {
-        return $this->innerJoinWith([
-            'locationTag' => fn (ActiveQuery $query) => $query->onCondition([LocationTag::tableName() . '.[[tag_id]]' => $tagId]),
-        ], $eagerLoading);
+        $onCondition = fn (ActiveQuery $query) => $query->onCondition([LocationTag::tableName() . '.[[tag_id]]' => $tagId]);
+
+        return $eagerLoading
+            ? $this->selectWith('locationTag', 'INNER JOIN', $onCondition)
+            : $this->innerJoinWith(['locationTag' => $onCondition], false);
     }
 }